initial setup

+++ vim-plug/plugins.vim
@@ -1,4 +1,4 @@
-" Auto-install vim-plug
+" Auto-install vim-plug if it's not already installed when starting neovim
 if empty(glob('~/.config/nvim/autoload/plug.vim'))
     silent !curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs
         \ https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
